% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFASTICAALG.R
\name{EMFASTICAALG}
\alias{EMFASTICAALG}
\title{EMFASTICAALG}
\usage{
EMFASTICAALG(DataMatrix, numCluster, h = 0, maxiter = 300, icaiter = 150,
  tol = 1e-06, verb = TRUE, combine = TRUE, seednum = 82196, ...)
}
\arguments{
\item{DataMatrix}{A matrix of which the rows are data entries. Its dimension
is \code{n} by \code{r}.}

\item{numCluster}{Predetermined number of mixing components \code{m}.}

\item{h}{Bandwidth. If \code{h} is set equal zero (default), iterative
bandwidth selection will be used.}

\item{maxiter}{Maximum number of iterations. Default is \code{300}.}

\item{icaiter}{Maximum number of ICA iterations in each step. Default is
\code{150}.}

\item{tol}{Threshold that defines convergence (of the outer loop). Default
is \code{1}\code{e}-\code{6}.}

\item{verb}{\code{TRUE} (default) or \code{FALSE} indicating whether to
print out info at each iteration.}

\item{combine}{\code{TRUE} (default) or \code{FALSE} indicating whether to
implement the ICA step.}

\item{seednum}{Seed number (default is \code{82196}) used in kmeans before
1st iteration.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use 
The returned value is an \code{EMFASTICAALG} object which consists of a list of items:
\item{$InputData }{A matrix of which the columns are data entries. Its
dimension is \code{r} by \code{n}.} \item{$Lambdas }{A matrix where rows
store estimated mixing weights from each iteration.} \item{$WMtrs }{List of
\code{r} by \code{r} unmixing matrices for each of the m clusters.}
\item{$WUnmixZ }{List of unmixing matrices for whitened data for each of the
m clusters.} \item{$OriginalSignals }{List of Recovered ICA components for
each of the \code{m} clusters.} \item{$ProductDensity }{\code{m} by \code{n}
matrix where each row stores the estimated density value of the observed
data points for each of the m clusters.} \item{$MembershipProbs }{\code{n}
by \code{m} matrix where each row stores the component membership
probabilities of the corresponding data point.} \item{$ObjValue }{Vector
holding values of data loglikelihood.} \item{$ICABandWidth }{Matrix holding
choices of bandwidth for original signals.} \item{$call }{The function call
that results in the returned object.} \item{$time }{Computing time elapsed
in second.} %% \item{comp2 }{Description of 'comp2'} %% ...
}
\description{
An R wrapper for carrying out NSMM-ICA on nonparametric multivariate ICA
mixture data.
}
\examples{

## An Example that runs the NSMM-ICA algorithm on Cohen's tone data
data(tonedata, package="mixtools")
b <- EMFASTICAALG(tonedata, 2, h=0, tol=1e-8)  

}
\keyword{~kwd1}
\keyword{~kwd2}
