\name{icamix-internal}
\alias{EMInterwovenFastICA}
\alias{WtsFastICA}
\alias{WtsKde}
\alias{restore_rng}
\alias{save_rng}

\title{Kernels for 'icamix' Functions}
\description{
  Internal kernel functions for the package \code{icamix}.
}
%\usage{
%EMInterwovenFastICA <- function(DataMtr, MembershipProbs_, bandWidth, maxIteration = 300L, icaIteration = 150L, tolerance = 1e-6, verbose = TRUE, combine = TRUE)
%WtsFastICA <- function(X, wts, Wconst, verbose = TRUE, alpha = 1, maxIteration = 60L, tolerance = 1e-6)
%WtsKde <- function(X, wts, grid, h)
%restore_rng <- function(savefile)
%save_rng <- function(savefile="tempfile1")
%}

%\arguments{...
%}

\details{
  These are usually not to be called by the user.
}

\keyword{internal}