\name{icamix-package}
\alias{icamix-package}
\alias{icamix}
\docType{package}
\title{
icamix: Estimation for ICA Mixture Model
}
\description{
The icamix package provides R functions as well as a C++ library which facilitate the estimation of ICA mixture models. 
}
\details{
\tabular{ll}{
Package: \tab icamix\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-30\cr
License: \tab GPL-2 | GPL-3\cr
}
We have developed the NSMM-ICA algorithm which is implemented in the main function EMFASTICAALG that currently integrates npEM and Fast-ICA for non-parametric estimation of ICA mixture models.
}
\author{
Xiaotian Zhu, David R. Hunter

Maintainer: Xiaotian Zhu <xxz131@psu.edu>
}
\references{
\href{http://sites.stat.psu.edu/~xxz131/XTZ_Dissertation.pdf}{http://sites.stat.psu.edu/~xxz131/XTZ_Dissertation.pdf}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
