% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFASTICAALG.R
\name{plot.summary.EMFASTICAALG}
\alias{plot.summary.EMFASTICAALG}
\title{plot.summary.EMFASTICAALG}
\usage{
\method{plot}{summary.EMFASTICAALG}(x, vec1 = c(1:2), vec2 = c(1:2), ...)
}
\arguments{
\item{x}{An \code{summary.EMFASTICAALG} object.}

\item{vec1}{An integer vector of length two specifying the coordinates with
respect to which the data is scatter plotted.}

\item{vec2}{An integer vector of length two specifying the coordinates with
respect to which the original signal for each mixture component is scatter
plotted.}
}
\description{
\code{plot} method for class \code{summary.EMFASTICAALG}.
}
\examples{

## An Example that runs the NSMM-ICA algorithm on Cohen's tone data
data(tonedata, package="mixtools")

b <- EMFASTICAALG(tonedata, 2, h=0, tol=1e-8)
plot(summary(b))


}
\keyword{~kwd1}
\keyword{~kwd2}
