% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFASTICAALG.R
\name{print.EMFASTICAALG}
\alias{print.EMFASTICAALG}
\title{print.EMFASTICAALG}
\usage{
\method{print}{EMFASTICAALG}(x, ...)
}
\arguments{
\item{x}{An \code{EMFASTICA} object.}
}
\description{
\code{\link[base]{print}} method for class \code{EMFASTICAALG}.
}
\examples{

## An Example that runs the NSMM-ICA algorithm on Cohen's tone data
data(tonedata, package="mixtools")

b <- EMFASTICAALG(tonedata, 2, h=0, tol=1e-8)
print(b)

}
\keyword{~kwd1}
\keyword{~kwd2}
