% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFASTICAALG.R
\name{print.summary.EMFASTICAALG}
\alias{print.summary.EMFASTICAALG}
\title{print.summary.EMFASTICAALG}
\usage{
\method{print}{summary.EMFASTICAALG}(x, ...)
}
\arguments{
\item{x}{An \code{summary.EMFASTICA} object.}
}
\description{
\code{\link[base]{print}} method for class \code{summary.EMFASTICAALG}.
}
\examples{

## An Example that runs the NSMM-ICA algorithm on Cohen's tone data
data(tonedata, package="mixtools")

b <- EMFASTICAALG(tonedata, 2, h=0, tol=1e-8)
print(summary(b))

}
\keyword{~kwd1}
\keyword{~kwd2}
