% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFASTICAALG.R
\name{summary.EMFASTICAALG}
\alias{summary.EMFASTICAALG}
\title{summary.EMFASTICAALG}
\usage{
\method{summary}{EMFASTICAALG}(object, ...)
}
\arguments{
\item{object}{An EMFASTICAALG object.}
}
\value{
The returned value is a "summary.EMFASTICAALG" object which consists
a list: %% ~Describe the value returned %% If it is a LIST, use
\item{$inputData }{A matrix of which the columns are data entries. Its
dimension is \code{r} by \code{n}.} \item{$originSig }{List of Recovered ICA
components for each of the \code{m} clusters.} \item{$call }{The function
call which results in the corresponding EMFASTICAALG object.} \item{$time
}{Computing time elaped in second.} \item{$numIter }{Total number of
iterations.} \item{$lastObj }{Objective function value from the last
iteration.} \item{$compMeans }{Means of each mixture component.}
\item{$compVars }{Covariances of each mixture component.} \item{$numObs
}{Total number of observations.} \item{$numAtr }{Dimension of data points.}
\item{$numCls }{Number of mixture components.} \item{$estWts }{Estimated
mixing weights.} \item{$estCls }{A factor whose levels represent estimated
class membership.} \item{$dataLklhd }{A vector storing data loglikelihood
from each iteration.} %% ...
}
\description{
\code{\link[base]{summary}} method for class \code{EMFASTICAALG}.
}
\examples{

## An Example that runs the NSMM-ICA algorithm on Cohen's tone data
data(tonedata, package="mixtools")

b <- EMFASTICAALG(tonedata, 2, h=0, tol=1e-8)
summary(b)

}
